refactor: consolidate shared benchmark utils into packages/shared/
- Move deriveTierGroupName and calculateTierDistributions to
  packages/shared/benchmark-utils.ts (single source of truth)
- Backend and frontend re-export from shared instead of duplicating
- Remove local implementations from report-routes.ts and
  frontend report-utils.ts
- Use Number.MAX_SAFE_INTEGER as sentinel for unranked tiers
- Remove unused athleteAge variable from getBenchmarkComparisons
