+++ packages/api/routes/report-routes.ts
@@ -36,6 +36,7 @@ import { RATE_LIMITS, RATE_LIMIT_WINDOW_MS } from "../constants/rate-limits";
 import { jsPDF } from "jspdf";
 import autoTable from "jspdf-autotable";
 import { isLowerBetter, sortAthletesByMetric, getBenchmarkLabel } from "../utils/report-utils";
+import { calculateTierDistributions } from "@shared/benchmark-utils";
 import { requireRole } from "../permissions/middleware";
 import { emailService } from "../services/email-service";
 import { getPushNotificationService, type NotificationPayload } from "../services/push-notification-service";
@@ -4417,62 +4418,6 @@ function calculateBenchmarkAchievements(athleteRankings: any[], teamStatistics?:
     });
 
   return achievements;
-}
-
-/**
- * Calculate tier distribution across athletes for team reports.
- * Returns how many athletes fall into each tier for each tier group.
- *
- * NOTE: Identical logic exists in packages/web/src/components/reports/report-utils.ts
- * for frontend rendering. Changes here must be mirrored there.
- */
-function calculateTierDistributions(athleteRankings: any[]): Array<{
-  metricCode: string;
-  tierGroupName: string;
-  tiers: Array<{ tierName: string; tierColor: string; tierOrder: number; count: number }>;
-}> {
-  if (!athleteRankings || athleteRankings.length === 0) return [];
-
-  // Map: "metricCode:tierGroupName" -> { tiers: Map<tierName, { color, order, count }> }
-  const groupMap = new Map<string, {
-    metricCode: string;
-    tierGroupName: string;
-    tiers: Map<string, { tierColor: string; tierOrder: number; count: number }>;
-  }>();
-
-  for (const athlete of athleteRankings) {
-    if (!athlete.benchmarkComparisons) continue;
-    for (const [metric, comparisons] of Object.entries(athlete.benchmarkComparisons) as [string, any[]][]) {
-      for (const comp of comparisons) {
-        if (!comp.tierName || !comp.tierGroupName) continue;
-        const key = `${metric}:${comp.tierGroupName}`;
-        if (!groupMap.has(key)) {
-          groupMap.set(key, {
-            metricCode: metric,
-            tierGroupName: comp.tierGroupName,
-            tiers: new Map(),
-          });
-        }
-        const group = groupMap.get(key)!;
-        if (!group.tiers.has(comp.tierName)) {
-          group.tiers.set(comp.tierName, {
-            tierColor: comp.tierColor || 'gray',
-            tierOrder: comp.tierOrder || 99,
-            count: 0,
-          });
-        }
-        group.tiers.get(comp.tierName)!.count++;
-      }
-    }
-  }
-
-  return Array.from(groupMap.values()).map(({ metricCode, tierGroupName, tiers }) => ({
-    metricCode,
-    tierGroupName,
-    tiers: Array.from(tiers.entries())
-      .map(([tierName, data]) => ({ tierName, ...data }))
-      .sort((a, b) => a.tierOrder - b.tierOrder),
-  }));
 }
 
 /**
